requirements.txt
flask
flask-cors
pymupdf
faiss-cpu
sentence-transformers
llama-cpp-python
numpy
fitz
crewai